Conflicts:
	KeWeiOMS.Domain/Entities/ProductType.cs
	KeWeiOMS.v11.suo

+++ KeWeiOMS.Domain/Entities/OrderPackRecordType.cs
@@ -58,12 +58,11 @@ namespace KeWeiOMS.Domain
         /// <summary>
         /// 包装系数
         /// </summary>
-        public virtual int PackCoefficient { get; set; }
+        public virtual double PackCoefficient { get; set; }
 
         /// <summary>
         /// 包装系数
         /// </summary>
         public virtual string SKU { get; set; }
-
     }
 }
